ui: minimal viable automated session creation
Opening webterm with a query parameter "into" containing a valid
session/agent name, will either select that session if it already
exists, or open a new session with that name that is connected to the
agent with the exact same name.

There are parts of this where we could opt into more complexity, like
separating the agent name from the session name, or being able to
specify agents that the handler should connect to, but whether that is
actually needed or useful in practice is not yet clear. The simple
version provided here will at least allow for cases like docket app
tiles linking to themselves in the web terminal, which should be a
direct improvement to frontend UX of TUI applications.