meta: deprecate @urbit/api
See: #176

This is a tangential task to an ongoing effort to implement Userspace Permissions UI (#182)

As a prerequisite for UP: the npm SDK which frontend developers use for integrating with Landscape APIs needs to be updated with interfaces for the new pokes and scries.

Previously, this would be `@urbit/api`. However, per this PR the package is now deprecated:
https://github.com/tloncorp/landscape/issues/176

Tlon has long had plans to release a successor to Indigo 1. The vision at this point is to create a new `@tloncorp/gear` package which will provide types for integrating with Landscape, as well as a component library and useful hooks for querying and managing state.

So why this PR: Since the new package will not have the Webterm types and utility functions, I'm migrating them from the deprecated package to this repo.